Combine both inputs into MessageInput, rename InputBase to MessageInputBase (again)
